requirements.txt
# Base networking and environment
requests>=2.31.0
python-dotenv>=1.0.1

# AI Frameworks
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0

# Data Validation & Logic
pydantic>=2.11.0
ratelimit>=2.2.1

# MCP (Model Context Protocol)
mcp>=1.2.0
fastmcp>=0.4.1

# PDF Processing (Generation & Reading)
fpdf2>=2.7.0
PyPDF2>=3.0.0